MAINT: Factors-out asset db schema
