+++ zipline/assets/asset_writer.py
@@ -27,6 +27,11 @@ import sqlalchemy as sa
 
 from zipline.errors import SidAssignmentError, AssetDBVersionError
 from zipline.assets._assets import Asset
+from zipline.assets.asset_db_schema import (
+    generate_asset_db_metadata,
+    asset_db_table_names,
+    ASSET_DB_VERSION,
+)
 
 SQLITE_MAX_VARIABLE_NUMBER = 999
 
@@ -164,12 +169,6 @@ def _generate_output_dataframe(data_subset, defaults):
     return output
 
 
-# Define a version number for the database generated by these writers
-# Increment this version number any time a breaking change is made to the
-# schema and readers of the database
-ASSET_DB_VERSION = 0
-
-
 def check_version_info(version_table, expected_version):
     """
     Checks for a version value in the version table.
@@ -213,143 +212,6 @@ def write_version_info(version_table, version_value):
 
     """
     sa.insert(version_table, values={'version': version_value}).execute()
-
-
-# A list of the names of all tables in the assets db
-asset_db_table_names = ['version_info', 'equities', 'futures_exchanges',
-                        'futures_root_symbols', 'futures_contracts',
-                        'asset_router']
-
-
-def _equities_table_schema(metadata):
-    return sa.Table(
-        'equities',
-        metadata,
-        sa.Column(
-            'sid',
-            sa.Integer,
-            unique=True,
-            nullable=False,
-            primary_key=True,
-        ),
-        sa.Column('symbol', sa.Text),
-        sa.Column('company_symbol', sa.Text, index=True),
-        sa.Column('share_class_symbol', sa.Text),
-        sa.Column('fuzzy_symbol', sa.Text, index=True),
-        sa.Column('asset_name', sa.Text),
-        sa.Column('start_date', sa.Integer, default=0, nullable=False),
-        sa.Column('end_date', sa.Integer, nullable=False),
-        sa.Column('first_traded', sa.Integer, nullable=False),
-        sa.Column('exchange', sa.Text),
-    )
-
-
-def _futures_exchanges_schema(metadata):
-    return sa.Table(
-        'futures_exchanges',
-        metadata,
-        sa.Column(
-            'exchange',
-            sa.Text,
-            unique=True,
-            nullable=False,
-            primary_key=True,
-        ),
-        sa.Column('timezone', sa.Text),
-    )
-
-
-def _futures_root_symbols_schema(metadata, futures_exchanges):
-    return sa.Table(
-        'futures_root_symbols',
-        metadata,
-        sa.Column(
-            'root_symbol',
-            sa.Text,
-            unique=True,
-            nullable=False,
-            primary_key=True,
-        ),
-        sa.Column('root_symbol_id', sa.Integer),
-        sa.Column('sector', sa.Text),
-        sa.Column('description', sa.Text),
-        sa.Column(
-            'exchange',
-            sa.Text,
-            sa.ForeignKey(futures_exchanges.c.exchange),
-        ),
-    )
-
-
-def _futures_contracts_schema(metadata, futures_root_symbols,
-                              futures_exchanges):
-    return sa.Table(
-        'futures_contracts',
-        metadata,
-        sa.Column(
-            'sid',
-            sa.Integer,
-            unique=True,
-            nullable=False,
-            primary_key=True,
-        ),
-        sa.Column('symbol', sa.Text, unique=True, index=True),
-        sa.Column(
-            'root_symbol',
-            sa.Text,
-            sa.ForeignKey(futures_root_symbols.c.root_symbol),
-            index=True
-        ),
-        sa.Column('asset_name', sa.Text),
-        sa.Column('start_date', sa.Integer, default=0, nullable=False),
-        sa.Column('end_date', sa.Integer, nullable=False),
-        sa.Column('first_traded', sa.Integer, nullable=False),
-        sa.Column(
-            'exchange',
-            sa.Text,
-            sa.ForeignKey(futures_exchanges.c.exchange),
-        ),
-        sa.Column('notice_date', sa.Integer, nullable=False),
-        sa.Column('expiration_date', sa.Integer, nullable=False),
-        sa.Column('auto_close_date', sa.Integer, nullable=False),
-        sa.Column('contract_multiplier', sa.Float),
-    )
-
-
-def _asset_router_schema(metadata):
-    return sa.Table(
-        'asset_router',
-        metadata,
-        sa.Column(
-            'sid',
-            sa.Integer,
-            unique=True,
-            nullable=False,
-            primary_key=True),
-        sa.Column('asset_type', sa.Text),
-    )
-
-
-def _version_table_schema(metadata):
-    return sa.Table(
-        'version_info',
-        metadata,
-        sa.Column(
-            'id',
-            sa.Integer,
-            unique=True,
-            nullable=False,
-            primary_key=True,
-        ),
-        sa.Column(
-            'version',
-            sa.Integer,
-            unique=True,
-            nullable=False,
-        ),
-        # This constraint ensures a single entry in this table
-        sa.CheckConstraint('id <= 1'),
-    )
 
 
 class AssetDBWriter(with_metaclass(ABCMeta)):
@@ -480,23 +342,16 @@ class AssetDBWriter(with_metaclass(ABCMeta)):
         constraints : bool, optional
             If True, create SQL ForeignKey and PrimaryKey constraints.
         """
-        metadata = sa.MetaData(bind=engine)
+        tables_already_exist = self.check_for_tables(engine)
+        metadata = generate_asset_db_metadata(bind=engine)
 
-        tables_already_exist = self.check_for_tables(engine=engine)
+        self.version_info = metadata.tables['version_info']
+        self.equities = metadata.tables['equities']
+        self.futures_exchanges = metadata.tables['futures_exchanges']
+        self.futures_root_symbols = metadata.tables['futures_root_symbols']
+        self.futures_contracts = metadata.tables['futures_contracts']
+        self.asset_router = metadata.tables['asset_router']
 
-        self.equities = _equities_table_schema(metadata)
-        self.futures_exchanges = _futures_exchanges_schema(metadata)
-        self.futures_root_symbols = _futures_root_symbols_schema(
-            metadata=metadata,
-            futures_exchanges=self.futures_exchanges,
-        )
-        self.futures_contracts = _futures_contracts_schema(
-            metadata=metadata,
-            futures_root_symbols=self.futures_root_symbols,
-            futures_exchanges=self.futures_exchanges,
-        )
-        self.asset_router = _asset_router_schema(metadata)
-        self.version_info = _version_table_schema(metadata)
         # Create the SQL tables if they do not already exist.
         metadata.create_all(checkfirst=True)
