+++ zipline/assets/assets.py
@@ -38,8 +38,6 @@ from zipline.assets import (
 )
 from zipline.assets.asset_writer import (
     split_delimited_symbol,
-    check_version_info,
-    ASSET_DB_VERSION,
     asset_db_table_names,
     SQLITE_MAX_VARIABLE_NUMBER
 )
@@ -105,9 +103,6 @@ class AssetFinder(object):
         metadata.reflect(only=asset_db_table_names)
         for table_name in asset_db_table_names:
             setattr(self, table_name, metadata.tables[table_name])
-
-        # Check the version info of the db for compatibility
-        check_version_info(self.version_info, ASSET_DB_VERSION)
 
         # Cache for lookup of assets by sid, the objects in the asset lookup
         # may be shared with the results from equity and future lookup caches.
